Change id type from INT to BIGINT

+++ src/main/resources/init.sql
@@ -1,15 +1,16 @@
+/*Aspect tables for jdbcTemplate*/
 DROP TABLE IF EXISTS accessed_price;
 CREATE TABLE accessed_price (
-  id INT PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   counter BIGINT
 );
 DROP TABLE IF EXISTS accessed_name;
 CREATE TABLE accessed_name (
-  id INT PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   counter BIGINT
 );
 DROP TABLE IF EXISTS tickets_bought;
 CREATE TABLE tickets_bought (
-  id INT PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   counter BIGINT
 );
